chore: Exclude .env and 'env' directory from version control
This commit updates the .gitignore file to exclude the .env file and 'env' directory from version control. The .env file contains environment-specific configurations and should not be tracked by Git. Additionally, the 'env' directory is excluded to manage environment-specific configurations and dependencies. These changes improve the cleanliness and efficiency of the repository.